fail count to long

+++ Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) throws InterruptedException {
         RateLimiter rateLimiter = new RateLimiter(100);
-        int failCount = 0;
+        long failCount = 0L;
         for(int i=0;i<200;i++){
             Thread.sleep(5);
             RequestResult result = rateLimiter.tryNewRequest();
